Improved max node id

+++ dpu/quicksort.c
@@ -30,7 +30,7 @@ uint32_t current_split = 0;
 MUTEX_INIT(splits_mutex);
 
 /*This is the main quicksort function. The sample will be moved from the current location to the top of the heap*/
-void sort_sample(uint32_t edges_in_sample, __mram_ptr edge_t* sample_from, edge_t* wram_buffer_ptr, int64_t max_node_id){
+void sort_sample(uint32_t edges_in_sample, __mram_ptr edge_t* sample_from, edge_t* wram_buffer_ptr, uint32_t max_node_id){
 
     //Number of edges per tasklet
     uint32_t nr_edges_tasklets = me() < edges_in_sample % NR_TASKLETS ? edges_in_sample/NR_TASKLETS + 1 : edges_in_sample/NR_TASKLETS;
